jarvis/requirements.txt
SpeechRecognition==3.10.0
pyttsx3==2.90
pyaudio==0.2.13
requests==2.31.0
beautifulsoup4==4.12.2
wikipedia==1.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pywhatkit==5.4
wolframalpha==5.0.0
